sim: Move management responsibility from probe listener
Listeners always relate to a specific probe points; thus,
their management should be done by the points, not from
the listeners themselves.

Explicitly call the probe manager to create new listeners,
and use smart pointers to manage them. This fixes memory
leaks within the probes of the mem dir.

As a side effect, renamed attachedProbePointList to listeners,
since it is a vector of probe listeners, not probe points.

As a side effect, removed the duplicated/unused vectors of
listener pointers.

Change-Id: Ib7d70cb51d346987ba4b0e7f969aa0274f19a3e4
Issued-on: https://gem5.atlassian.net/browse/GEM5-857
Signed-off-by: Daniel R. Carvalho <odanrc@yahoo.com.br>