arch-riscv: Implement Zcmt
Implement table jump (cm.jt and cm.jalt) instructions

According to spec, when jump to the JVT entry address. We need to take
two fetch steps. The first is fetching the zcmt instruction. The second
is fetching the JVT entry. The JVT entries is considered the extension
of table jump instructions. The JVT entries read from icache_port and
enable get them in I-cache, so we can't follow the generic way to
load JVT entry from memory helper.

In gem5 implementation we will get two StaticInst for cm.jalt. The first
is move the PC to address of JVT entry, and mark the pc_state to zcmt
fetch mode. When in zcmt mode, we will get the StaticInst that jumps to
the JVT entry, and leaves the zcmt fetch mode.

If there is fault during second fetching, xEPC is set to instruction
address, xCAUSE is set to expect type of fault, xTVAL is set to address
of JVT entry

Spec: https://github.com/riscv/riscv-isa-manual/blob/main/src/zc.adoc#zcmt

Bug: 322083446
Test: None
Change-Id: I8e0caca35a22a51386677d70fa09c3dc97e4610a