base: Make BaseGdbRegCache::data() non constant
The method is defined as const but the caller will actually modify the
content of the structure directly with the pointer in
BaseRemoteGDB::cmdRegW. The member access in the const method are
actually treated as const and will cause error if we use
reinterpret_cast instead.

Remove the const tag to align the expectation of the virtual method.