arch-arm: Do not compute purifyTaggedAddr in checkPermissions
purifyTaggedAddr is known to be an expensive computation
regardless of the memoization we do, as it sits in the
critical path from a host performance point of view
(instruction fetch).
In checkPermissions64 we compute it without really needing
the tag purification. The only place where it is used
is to check for PCAlignment, but the alignment checks the
3LSBs whereas a potential tag would be stored in the most
significant ones

Change-Id: I9f39db658c3575dcbacb5351813ff9bb3775046d
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>