arch, cpu: Add generic getValidAddr to correct exetrace symbol table
The getValidAddr is the method get virtual address with valid bits.
It is useful to get the correct symbol table via valid virtual address.

For ARM, we have `purifyTaggedAddr` to get the right virtual address.
For RISC-V, we only get lower 32 bits in RV32 mode to get the right
symbol table.

Change-Id: I33ad7bec6e7ea4ec82cb1b3a7f521432c6d735b6