tests: add non KVM CPU type switch tests
This commit adds tests for non KVM type switching. The CPU
switches are atomic to timing, atomic to minor, atomic to o3,
timing to minor, and timing to o3. There are at least 3 tests
for each type of CPU switch, and the original plan was to only have
3 tests for each type, to avoid adding too many tests. Each set
has one test for each ISA out of X86, ARM, and RISCV, and uses 1,
2, or 4 cores. Two tests of each set use the classic memory system
and the remaining test uses MESI two level. I tried to shuffle
around the combinations of ISAs, cores, and memory system to
maximize the range of what was tested without using too many tests.