arch-arm: Print the ESR_ELx register when generating a fault
Current debug print for an AArch64 fault is based on the AArch32 model,
where an exception vector holds the entry point to different types of
exception handlers. As such, the new PC is relevant in extracting
information about the type of the fault.
This is less true in AArch64 as most exceptions share the same entry
point. It is far more interesting when tracing your faults to gather the
motivation behind a generated fault.  This is encoded in the exception
syndrome register (ESR).  We therefore replace the newVec entry with the
ESR value

Change-Id: I18a09aa73682e3303b538d35df55a3460c4d8563
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>