arch-arm: Provide trapping helper for _EL12 registers
The trapping logic for EL_12 registers is mainly the following:

1 check if feature is present and have undefined instruction otherwise
2 check if we are in host mode and have undefined instruction otherwise
3 apply EL2 trapping logic

We can therefore elide most of the callbacks and rely on a template
instead which satisfies conditions above. The only difference is
in the order between 1 and 2. But as both should raise undefined
instruction exception, this does not really matter

Change-Id: If1fa74ec0916e065409baef5619c3727534d7728
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>