cpu: Incorrect BP update for atomic core
In the atomic core the branch predictor was not correctly
updated upon misprediction. On misprediction the history
was only squashed but not commited.
It has only a small impact on the accuracy because most of
the time a subsequent correctly predicted branch calls
the update function and both branches will be commited.
However, in case of two subsequent mispredictions the
earlier one which is already corrected with be squashed
again and the update is lost.

Signed-off-by: David Schall <david.schall@tum.de>