mem-ruby: Fix functional access in MI_example
In MI_example, when in MI state the block "Maybe_Stale" as in this
controller may have the most up to date value or it could be in the
network. For MII it is guaranteed that this controller has the most up
to date value because it received a PUTX_NACK.

This fixes one of the daily test failures.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>