mem-ruby: Fix missing RubySystem in PerfectCacheMemory's entries
MOESI_CMP_directory protocol crashes with one of the several
assertions in NetDest.cc. It happens because the entry type used to
instantiate a PerfectCacheMemory object in
MOESI_CMP_directory-L2cache.sm contains a NetDest object, so it
requires a RubySystem object to be manually set for it.

Instead of just receiving the block size, change PerfectCacheMemory to
receive a RubySystem object and use it to set the block size and call
ENTRY::setRubySystem if the entries require it.