mem-ruby: set RubySystem pointer during TBE alloc
Currently the RubySystem pointer is set when set_tbe is performed, which
effectively clears the NetDest objects from the TBE (if any). This is fine if
the TBE has been just allocated before set_tbe is called (no NetDest
info in the TBE). However, the CHI protocol has an action (RestoreFromHazard)
that performs a set_tbe over a TBE that had already been set, i.e., it
already has valid NetDest data.

This patch sets the RubySystem pointer when the TBE is allocated, which
is more natural and follows the style already adopted in the
PerfectCacheMemory class (#1864).