stdlib: Add a gem5 orchestrator exit
The gem5 orchestrator is a starting point and example of how the new
external process handler interface can be used. The orchestrator is a
bi-directional communication framework so that an external process can
both send commands to a running gem5 instance and receive responses.

The orchestrator sends commands using the externalProcessHandler in
gem5. The commands are simple json packages using the hypercall number
1000 (now reserved for this purpose). In this json message the
orchestrator can send a string function to run and a socket path to
communicate back (note that this only supports unix socket communication
right now, so likely no Windows support. This has not been tested on
MacOS.)

Two functions are initially supported:
- status: Returns the number of instructions and ticks executed so far
  and the workload that is running.
- get_stats: Returns a json object with all of the current statistic
  values.

This patch also includes a simple script that leverages the transimitter
script to run these two functions.

Note: The orchestrator will only support connecting to gem5 when its
running using the standard library. fs/se.py or other fully custom
scripts will not be supported.

Signed-off-by: Jason Lowe-Power <jason@lowepower.com>