configs: Add BadAddr responder to the L2Cache XBar
I was running a a test binary that had a load after a conditional branch to
address 0-immediate, i.e. 0xfff... which was working fine until I turned on
an l2 cache. Then I was hitting`fatal("Unable to find destination for ...`
which turns out to be because the L2XBar does not have a default
destination.

Change-Id: I7ff624b0ef058789bd9b0ea31dbd46189769bce4